requirements.txt
asgiref==3.5.2
Django==3.2.15
django-bootstrap5==22.2
django-ckeditor==6.5.1
django-environ==0.9.0
django-js-asset==2.0.0
django-taggit==3.1.0
gitdb==4.0.10
GitPython==3.1.30
Pillow==9.3.0
pip==23.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.6
setuptools==65.6.3
smmap==5.0.0
sqlparse==0.4.3
typing_extensions==4.4.0
wheel==0.38.4
